+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Query:** Raise `TypeError` when all values are non-numeric in given field for `min()`, `max()`, `avg()`
 - Fixed always `'None'` for wrong-parsed string values
 - Fixed replacing all used `node` and `relation` words in node / relation definitions
+- **Query:** Fixed `O(n^2)` for `exclude()` and `intersect()`
 
 ### Removed